+++ src/Field.php
@@ -29,6 +29,7 @@ use craft\models\CategoryGroup;
 use craft\models\ImageTransform;
 use craft\models\Section;
 use craft\models\Volume;
+use craft\services\ElementSources;
 use craft\web\View;
 use HTMLPurifier_Config;
 use HTMLPurifier_HTMLDefinition;
@@ -658,6 +659,7 @@ JS,
                 'elementType' => Category::class,
                 'refHandle' => Category::refHandle(),
                 'sources' => $categorySources,
+                'criteria' => ['uri' => ':notempty:'],
             ];
         }
 
@@ -722,12 +724,20 @@ JS,
             }
         }
 
+        $sources = array_values(array_unique($sources));
+
         if ($showSingles) {
             array_unshift($sources, 'singles');
         }
 
         if (!empty($sources)) {
             array_unshift($sources, '*');
+        }
+
+        // include custom sources
+        $customSources = $this->_getCustomSources(Entry::class);
+        if (!empty($customSources)) {
+            $sources = array_merge($sources, $customSources);
         }
 
         return $sources;
@@ -745,11 +755,19 @@ JS,
             return [];
         }
 
-        return Collection::make(Craft::$app->getCategories()->getAllGroups())
+        $sources = Collection::make(Craft::$app->getCategories()->getAllGroups())
             ->filter(fn(CategoryGroup $group) => $group->getSiteSettings()[$element->siteId]?->hasUrls ?? false)
             ->map(fn(CategoryGroup $group) => "group:$group->uid")
             ->values()
             ->all();
+
+        // include custom sources
+        $customSources = $this->_getCustomSources(Category::class);
+        if (!empty($customSources)) {
+            $sources = array_merge($sources, $customSources);
+        }
+
+        return $sources;
     }
 
     /**
@@ -774,10 +792,37 @@ JS,
             $volumes = $volumes->filter(fn(Volume $volume) => $userService->checkPermission("viewAssets:$volume->uid"));
         }
 
-        return $volumes
+        $sources = $volumes
             ->map(fn(Volume $volume) => "volume:$volume->uid")
             ->values()
             ->all();
+
+        // include custom sources
+        $customSources = $this->_getCustomSources(Asset::class);
+        if (!empty($customSources)) {
+            $sources = array_merge($sources, $customSources);
+        }
+
+        return $sources;
+    }
+
+    /**
+     * Returns custom element sources keys for given element type.
+     *
+     * @param string $elementType
+     * @return array
+     */
+    private function _getCustomSources(string $elementType): array
+    {
+        $customSources = [];
+        $elementSources = Craft::$app->getElementSources()->getSources($elementType, 'modal');
+        foreach ($elementSources as $elementSource) {
+            if ($elementSource['type'] === ElementSources::TYPE_CUSTOM && isset($elementSource['key'])) {
+                $customSources[] = $elementSource['key'];
+            }
+        }
+
+        return $customSources;
     }
 
     /**